// packages/frontend/src/generated/types/PhysicalTransferResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { DocumentStatus } from './DocumentStatus.ts'
import type { PhysicalTransferItemResponse } from './PhysicalTransferItemResponse.ts'

/**
 * @description Response DTO for the `physical_storage_transfer` entity.
 */
export interface PhysicalTransferResponse {
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string
   */
  date: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type string
   */
  endCargoOps: string
  /**
   * @type string,null
   */
  executedAt?: string | null
  /**
   * @type string,null, uuid
   */
  executedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type array
   */
  items: PhysicalTransferItemResponse[]
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string,null
   */
  revertedAt?: string | null
  /**
   * @type string,null, uuid
   */
  revertedBy?: string | null
  /**
   * @type string
   */
  startCargoOps: string
  /**
   * @type string
   */
  status: DocumentStatus
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}